requirements.txt
locust
azure-storage-blob
azure-cosmosdb-table
click
Flask
itsdangerous
Jinja2
MarkupSafe
Werkzeug
